% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prized_goat.R
\name{prized_goat}
\alias{prized_goat}
\title{A tragic function that I do not know how to get to work and doesnt
follow the assignment, but I am trying my best :)}
\usage{
prized_goat(K, r)
}
\arguments{
\item{K}{number of goat score 1}

\item{r}{number of goat score 2}
}
\value{
the total score for got in this imaginary competition
}
\description{
A tragic function that I do not know how to get to work and doesnt
follow the assignment, but I am trying my best :)
}
